💩 wip: Created the seed script
